Update doom.py with Milestone 3

+++ doom.py
@@ -5,7 +5,8 @@ import datetime
 import time
 import sqlite3
 import signal
-import pynput
+import pynput #type:ignore
+import threading
 
 app = Flask(__name__)
 
@@ -16,14 +17,20 @@ cur = mydb.cursor() #creates cursor in database to execute commands
 user: {username: string[32], password: string[32]}
 current: {username: string[32], password: string[32]}
 active_timer: {id: int, duration: int, remaining time: string(datetime), date: string(datetime)}
+break: {id int, break_start int, break_duration int}
+distraction: {id: int, duration: int, end_time: string(datetime), cause: string}
 completed_timers: {id: int, duration: int, focus%: float, date: String(ISOFormat)}
 goals: {id: int, username string{32}, start date: string(date), end date: string(date), target_focus_percentage: int}
 '''
 cur.execute("CREATE TABLE IF NOT EXISTS user(username VARCHAR[32] NOT NULL PRIMARY KEY, password VARCHAR[32] NOT NULL)")
 cur.execute("CREATE TABLE IF NOT EXISTS current(username VARCHAR[32] NOT NULL, password VARCHAR[32] NOT NULL)")
 cur.execute("CREATE TABLE IF NOT EXISTS active_timer(id INTEGER PRIMARY KEY AUTOINCREMENT, duration int NOT NULL, remaining_time VARCHAR[32] NOT NULL, date VARCHAR[32] NOT NULL)")
+cur.execute("CREATE TABLE IF NOT EXISTS break(id INTEGER PRIMARY KEY AUTOINCREMENT, break_start int NOT NULL, break_duration int NOT NULL)")
+cur.execute("CREATE TABLE IF NOT EXISTS distraction(id INTEGER PRIMARY KEY AUTOINCREMENT, duration int NOT NULL, end_time VARCHAR[32] NOT NULL, cause VARCHAR[32] NOT NULL)")
 cur.execute("CREATE TABLE IF NOT EXISTS completed_timers(id INTEGER PRIMARY KEY AUTOINCREMENT, duration int NOT NULL, focus float(24) NOT NULL, date VARCHAR[32] NOT NULL)")
 cur.execute("CREATE TABLE IF NOT EXISTS goals(id INTEGER PRIMARY KEY AUTOINCREMENT, username TEXT NOT NULL, start_date TEXT NOT NULL, end_date TEXT NOT NULL, target_focus_percentage INTEGER NOT NULL)")
+cur.execute("DELETE FROM active_timer")
+cur.execute("DELETE FROM distraction")
 res = cur.execute(f"SELECT username, password FROM current")
 #print(f"{res.fetchone()!r}")
 if f"{res.fetchone()!r}" == "None": #if current is empty (should only happen on first startup)
@@ -35,36 +42,67 @@ mydb.close() #closes the database thread
 #################################################################
 # START User_Input
 #################################################################
-def inputListener():
-    def receiveSignal(signalNumber, frame): #detects when SIGILL (crtl+C) is triggered
-        print('Received:', signalNumber)
-        pynput.keyboard.Listener.stop(listenerK) #stops listenerK
-        listenerK.join() #ends listenerK
-        pynput.mouse.Listener.stop(listenerM) #stops listenerM
-        listenerM.join() #ends listenerM
-        os.kill(os.getpid(), signal.SIGTERM) #kills Flask
-        return
-    
-    signal.signal(signal.SIGILL, receiveSignal) #recieves Kill Signal
-    signal.signal(signal.SIGINT, receiveSignal) #recieves Int_type Signal
+def receiveSignal(signalNumber, frame): #detects when SIGILL (crtl+C) is triggered
+    print('Received:', signalNumber)
+    global threadEnd
+    threadEnd = True #setting this will end the thread
+    inputMonitor.join() #ends inputMonitor
+    pynput.keyboard.Listener.stop(listenerK) #stops listenerK
+    listenerK.join() #ends listenerK
+    pynput.mouse.Listener.stop(listenerM) #stops listenerM
+    listenerM.join() #ends listenerM
+    os.kill(os.getpid(), signal.SIGTERM) #kills Flask
+    return
 
-    def on_key_press(key):
-        try:
-            print('alphanumeric key {0} pressed'.format(key.char))
-        except AttributeError:
-            print('special key {0} pressed'.format(key))
-    listenerK = pynput.keyboard.Listener(on_press=on_key_press) #creates listenerK to listen for key presses
-    listenerK.start()
+signal.signal(signal.SIGILL, receiveSignal)
+signal.signal(signal.SIGINT, receiveSignal)
 
-    def on_mouse():
-        print("Mouse")
-    listenerM = pynput.mouse.Listener(
-        on_move=on_mouse,
-        on_click=on_mouse,
-        on_scroll=on_mouse) #creates listenerK to listen for mouse actions (left/right/middle-click, movement, scrolling)
-    listenerM.start()
 
-inputListener()
+tester = "init" #use current render_template to change page without reload (DOESN'T WORK)
+threadEnd = False #flag to end inputMonitor thread
+durAFK = 0.0 #duration without input, in seconds
+
+def getAFK(): #spins in place, detecting if user has gone AFK
+    global tester
+    global threadEnd
+    global durAFK
+    while not threadEnd:
+        if durAFK > 5*60: #if 5 minutes have passed without input, [trigger AFK]
+            tester = "AFK"
+            #print("AFK")
+        #else:
+        durAFK += 0.1
+        time.sleep(0.1)
+    return
+
+def on_key_press(key):
+    '''try:
+        print('alphanumeric key {0} pressed'.format(key.char))
+    except AttributeError:
+        print('special key {0} pressed'.format(key))'''
+    global durAFK
+    durAFK = 0.0 #reset to 0 when key pressed
+listenerK = pynput.keyboard.Listener(on_press=on_key_press)
+listenerK.start()
+
+def on_mouse():
+    #print("Mouse")
+    global durAFK
+    durAFK = 0.0 #reset to 0 when mouse manipulated (LClick, RC, MC, Scroll, Move)
+listenerM = pynput.mouse.Listener(
+    on_move=on_mouse,
+    on_click=on_mouse,
+    on_scroll=on_mouse)
+listenerM.start()
+
+inputMonitor = threading.Thread(target=getAFK)
+inputMonitor.start() #starts inputMonitor as thread with function getAFK()
+
+@app.route('/process-data', methods=['POST']) #not viewable webpage, just used for fetch requests
+def process_data():
+    global durAFK
+    return str(durAFK > 5.0*60)
+
 #################################################################
 # END User_Input
 #################################################################
@@ -96,8 +134,9 @@ def focus():
     timerText = updateTime()
     goalText = setGoal()
     completeTimers = setComplete()
+    distText = setDistraction()
     
-    return render_template('focus.html', name=username, curTimer=timerText, comTimers=completeTimers, curGoal=goalText)
+    return render_template('focus.html', name=username, curTimer=timerText, comTimers=completeTimers, curGoal=goalText, curDist=distText)
 
 @app.route('/focus', methods=["POST"]) #executes on form submission
 def focusAdd():
@@ -109,11 +148,16 @@ def focusAdd():
     timerText = ""
     if str(request.form.get("tdur")) != "None": #if Focus form was submitted
         timerDur = str(request.form["tdur"]).strip() #duration of timer
-        
+        cur.execute("DELETE FROM break") #removes current break if overwriting timer
+        if str(request.form.get("bstr")) != "None": #if break exists
+            cur.execute(f'''INSERT INTO break(break_start, break_duration) VALUES (
+                        {request.form["bstr"]},
+                        {request.form["bdur"]})''')
+
         duration = int(timerDur)
         timeStart = str(datetime.datetime.now()) #start time of timer
         
-        res = cur.execute(f"SELECT id FROM active_timer") #fetch data from current focus timer
+        res = cur.execute(f"SELECT id FROM active_timer") #fetch data from current active_timer
         if str(res.fetchone()) != "None": #if current active_timer timer exists, then delete it
             cur.execute(f"DELETE FROM active_timer")
         cur.execute(f'''INSERT INTO active_timer(duration, remaining_time, date) VALUES (
@@ -121,6 +165,8 @@ def focusAdd():
                     {("0"+str(datetime.timedelta(minutes=int(duration/60), seconds=duration%60)))!r},
                     {timeStart!r})''') #inserts data about active_timer timer into active_timer table
         
+        cur.execute(f"DELETE FROM distraction") #empties distraction table
+
         #(pulled from updateTime())
         res = cur.execute(f"SELECT id, duration, remaining_time, date FROM active_timer") #load from active_timer table
         name = res.fetchone()
@@ -129,8 +175,8 @@ def focusAdd():
         remTime = 3600*remTime.hour + 60*remTime.minute + remTime.second #converts into total seconds
         timerText = f"{name[0]},{name[1]},{remTime},{name[3]}"
 
-    else: #if not focus form just submitted
-        timerText = updateTime() #update time in it (only needed to do when not focus form submitted)
+    else: #if not timer form just submitted
+        timerText = updateTime() #update time in it (only needed to do when not timer form submitted)
         
         if str(request.form.get("start_date")) != "None": #if Goal form was submitted
             goalBeg = str(request.form["start_date"]).strip()
@@ -141,16 +187,26 @@ def focusAdd():
                         {goalBeg!r},
                         {goalEnd!r},
                         {goalFoc})''')
+        
+        elif str(request.form.get("dname")) != "None": #if Distraction form was submitted
+            distName = str(request.form["dname"]).strip()
+            distDur = str(request.form["ddur"]).strip()
+            cur.execute(f'''INSERT INTO distraction(duration, cause, end_time) VALUES (
+                        {distDur!r},
+                        {distName!r},
+                        {str(datetime.datetime.now())!r})''')
     
     goalText = setGoal()
     completeTimers = setComplete()
+    distText = setDistraction()
     
     mydb.commit()
     mydb.close()
-    return render_template('focus.html', name=username, curTimer=timerText, comTimers=completeTimers, curGoal=goalText, redir=1)
+    return render_template('focus.html', name=username, curTimer=timerText, comTimers=completeTimers, curGoal=goalText, curDist=distText, redir=1)
 
-@app.route('/focusComplete') #executes on timer expiration
-def focusComplete():
+@app.route('/focusComplete/<forceStop>') #executes on timer expiration, or forcestop
+def focusComplete(forceStop):
+    print(forceStop)
     username = getUsername()
 
     mydb = sqlite3.connect("doom.db")
@@ -159,24 +215,77 @@ def focusComplete():
     res = cur.execute(f"SELECT id, duration, remaining_time, date FROM active_timer") #load active_timer timer from active_timer table
     name = res.fetchone()
 
-    timerDurs = name[1]
-
     startTime = datetime.datetime.fromisoformat(name[3])
-    timerDuration = timerDurs
+    timerDuration = name[1]
+
+    timerDurs = []
+    res = cur.execute(f"SELECT break_start, break_duration FROM break") #gets break_start and break_duration from break
+    names = res.fetchall()
+    if len(names) > 0: #if break exists
+        name = names[0]
+        timerDurs.append(name[0]) #first is time from timer_start to break_start, equal to break_start
+        timerDurs.append(name[1]) #second is duration from break_start to break end, eqaul to break_duration
+        timerDurs.append(timerDuration - name[0] - name[1]) #final is duration from break end to timer end, equal to timer_duration - (break_start + break_duration)
+    else:
+        timerDurs.append(timerDuration)
+
+    def checkOverlap(focusStart, focusDur, distEnd, distDur): #gets overlap between a focus timer and a distraction
+        if forceStop == "true":
+            focusEnd = datetime.datetime.now()
+        else:
+            focusEnd = focusStart + datetime.timedelta(minutes=int(focusDur/60), seconds=focusDur%60) #creates end datetime of focus timer
+        distStart = distEnd - datetime.timedelta(minutes=int(distDur/60), seconds=distDur%60) #creates start datetime of distraction
+        
+        #clamp distraction within focus
+        if distStart < focusStart:
+            distStart = focusStart
+        if distEnd > focusEnd:
+            distEnd = focusEnd
+        if distStart > focusEnd: #also clamps start of distraction by end of timer, only needed on forcestop
+            distStart = focusEnd
+        
+        if distStart < distEnd: #clamping doesn't affect other bound, so if whole dist outside focus, then this fails
+            return (distEnd - distStart).total_seconds() #return overlap betewen focus and distraction
+        return 0 #if no overlap, return 0
+    
+    timeDistracted = 0.0 #total time distracted
+    
+    res = cur.execute("SELECT duration, end_time FROM distraction")
+    for name in res.fetchall():
+        #distraction += f"{name[0]},{name[1]},{name[2]},{name[3]};"
+        startFocus = 0 #start of each focus timer, past startTime
+        for b in range(0, len(timerDurs), 2):
+            timeDistracted += checkOverlap(
+                startTime + datetime.timedelta(minutes=int(startFocus/60), seconds=startFocus%60),
+                int(timerDurs[b]),
+                datetime.datetime.fromisoformat(name[1]),
+                int(name[0])) #increments timeDistracted by overlap between current Focus and Distraction
+            if b+1 < len(timerDurs): #if not at final focus timer
+                startFocus += int(timerDurs[b]) + int(timerDurs[b+1]) #increments startFocus by current Focus and next Break durations
+            else:
+                startFocus += int(timerDurs[b]) #increments startFocus by current/final Focus
+
+    #print("timerDuration:",timerDuration)
+    #print("timerDuration:",str(datetime.datetime.now() - startTime))
+    if forceStop == "true":
+        timerDuration = (datetime.datetime.now() - startTime).seconds
 
     cur.execute(f'''INSERT INTO completed_timers(duration, focus, date) VALUES (
                 {timerDuration},
-                {100.0},
+                {100 * (1 - (timeDistracted/timerDuration)):.2f},
                 {str(startTime)!r})''')
     
     cur.execute(f"DELETE FROM active_timer")
-    
+    cur.execute(f"DELETE FROM distraction")
+    cur.execute(f"DELETE FROM break")
+
     goalText = setGoal()
     completeTimers = setComplete()
+    distText = setDistraction()
     
     mydb.commit()
     mydb.close()
-    return render_template('focus.html', name=username, curTimer="", comTimers=completeTimers, curGoal=goalText, redir=1)
+    return render_template('focus.html', name=username, curTimer="", comTimers=completeTimers, curGoal=goalText, curDist=distText, redir=1)
 
 @app.route('/focusRemove/<elementType>/<elementID>') #remove specified element (elementType = "timer" or "goal")
 def focusRemove(elementType, elementID):
@@ -185,7 +294,7 @@ def focusRemove(elementType, elementID):
     mydb = sqlite3.connect("doom.db")
     cur = mydb.cursor()
     
-    if elementType == "timer": #if Focus form was submitted
+    if elementType == "timer": #if timer form was submitted
         cur.execute(f"DELETE FROM completed_timers WHERE id={elementID}") #delete timer with elementName
     elif elementType == "goal":
         cur.execute(f"DELETE FROM goals WHERE id={elementID}") #delete goal with elementName
@@ -193,10 +302,11 @@ def focusRemove(elementType, elementID):
     timerText = updateTime() #update time in it
     goalText = setGoal()
     completeTimers = setComplete()
+    distText = setDistraction()
 
     mydb.commit()
     mydb.close()
-    return render_template('focus.html', name=username, curTimer=timerText, comTimers=completeTimers, curGoal=goalText, redir=1)
+    return render_template('focus.html', name=username, curTimer=timerText, comTimers=completeTimers, curGoal=goalText, curDist=distText, redir=1)
 
 def updateTime(): #updates time in active_timers table, and returns timerText to give to js
     mydb = sqlite3.connect("doom.db")
@@ -204,7 +314,7 @@ def updateTime(): #updates time in active_timers table, and returns timerText to
 
     res = cur.execute(f"SELECT id FROM active_timer") #fetch data from current active_timer
     if str(res.fetchone()) == "None": #if current active_timer is empty, then return ""
-        mydb.commit()
+        #mydb.commit()
         mydb.close()
         return ""
     
@@ -213,10 +323,10 @@ def updateTime(): #updates time in active_timers table, and returns timerText to
 
     duration = name[0]
     timeStart = name[1]
+    
     timeDiff = datetime.datetime.now() - datetime.datetime.fromisoformat(timeStart) #difference between now and timeStart
     timeDur = datetime.timedelta(seconds=duration)
     cur.execute(f"UPDATE active_timer SET remaining_time = '0{timeDur - timeDiff}' ") #updates active_timer's remaining time
-
 
     res = cur.execute(f"SELECT id, duration, remaining_time, date FROM active_timer") #load from active_timer table
     name = res.fetchone()
@@ -236,7 +346,7 @@ def setComplete():
     completeTimers = ""
     for name in res.fetchall():
         completeTimers += f"{name[0]},{name[1]},{name[2]},{name[3]};"
-    mydb.commit()
+    #mydb.commit()
     mydb.close()
     return completeTimers
 
@@ -247,9 +357,20 @@ def setGoal():
     goalText = ""
     for name in res.fetchall():
         goalText += f"{name[0]},{name[1]},{name[2]},{name[3]};"
-    mydb.commit()
+    #mydb.commit()
     mydb.close()
-    return goalText    
+    return goalText
+
+def setDistraction(): #DISTRACTION
+    mydb = sqlite3.connect("doom.db")
+    cur = mydb.cursor()
+    res = cur.execute("SELECT id, duration, end_time, cause FROM distraction")
+    distText = ""
+    for name in res.fetchall():
+        distText += f"{name[0]},{name[1]},{name[2]},{name[3]};"
+    #mydb.commit()
+    mydb.close()
+    return distText
 
 @app.route('/pwreset')
 def pwReset():
@@ -395,3 +516,17 @@ def error(e):
 
 if __name__ == '__main__':
     app.run(host='0.0.0.0', port=15271, debug=True)
+
+'''
+To-Do:
+    Change all file accessing to database accessing:
+        [X] Focus
+        [X] Complete
+        [X] Goal
+    Comment out stuff related to later Milestones:
+        Break
+        Distraction
+        AFK (not included yet)
+        Stats (not included yet)
+[X] Add button to remove focus goal
+'''
